Crate gcrypt

Initialization

The library must be initialized using gcrypt::init or gcrypt::init_fips_mode before using any other function in the library or wrapper. More information on initialization can be found in the libgcrypt documentation.

An example:

let token = gcrypt::init(|mut x| {
    x.disable_secmem();
});

The token returned by init is used as an argument to various functions in the library to ensure that initialization has been completed.

Reexports

pub extern crate gpg_error as error;

Modules

cipher

digest

kdf

mac

mpi

pkey

rand

sexp

Structs

Buffer

Error
A type wrapping errors produced by GPG libraries.

Initializer

Token

Traits

Wrapper

Functions

enable_memory_guard

get_token

init

init_fips_mode

is_initialized

Type Definitions

Result